+++ findex/src/main/java/com/example/findex/mapper/IndexDataMapper.java
@@ -1,6 +1,6 @@
 package com.example.findex.mapper;
 
-import com.example.findex.dto.data.IndexDataDto;
+import com.example.findex.dto.indexdata.data.IndexDataDto;
 import com.example.findex.entity.IndexData;
 import com.example.findex.entity.IndexInfo;
 import com.example.findex.entity.SourceType;
